Split radarr/sonarr files for cfsync
